Release 0.1

+++ src/deposit.h
@@ -3,7 +3,7 @@
 
 int check_input(int sum, int time);
 
-float calculation(int sum, int time);
+double calculation(int sum, int time);
 
 int income(int first_sum, int last_sum);
